+++ lib/hanami/action/params.rb
@@ -18,7 +18,7 @@ module Hanami
 
       # This is a Hanami::Validations extension point
       #
-      # @since x.x.x
+      # @since 0.7.0
       # @api private
       def self._base_rules
         lambda do
@@ -54,10 +54,51 @@ module Hanami
         @input
       end
 
+      # Returns structured error messages
+      #
+      # @return [Hash]
+      #
+      # @since 0.7.0
+      #
+      # @example
+      #   params.errors
+      #     # => {:email=>["is missing", "is in invalid format"], :name=>["is missing"], :tos=>["is missing"], :age=>["is missing"], :address=>["is missing"]}
       def errors
         @result.messages
       end
 
+      # Returns flat collection of full error messages
+      #
+      # @return [Array]
+      #
+      # @since 0.7.0
+      #
+      # @example
+      #   params.error_messages
+      #     # => ["Email is missing", "Email is in invalid format", "Name is missing", "Tos is missing", "Age is missing", "Address is missing"]
+      def error_messages(error_set = errors)
+        error_set.each_with_object([]) do |(key, messages), result|
+          k = Utils::String.new(key).titleize
+
+          _messages = if messages.is_a?(Hash)
+            error_messages(messages)
+          else
+            messages.map { |message| "#{k} #{message}" }
+          end
+
+          result.concat(_messages)
+        end
+      end
+
+      # Returns true if no validation errors are found,
+      # false otherwise.
+      #
+      # @return [TrueClass, FalseClass]
+      #
+      # @since 0.7.0
+      #
+      # @example
+      #   params.valid? # => true
       def valid?
         @result.success?
       end
@@ -74,31 +115,8 @@ module Hanami
 
       private
 
-      # @since x.x.x
-      # @api private
-      def _extract_params
-        # FIXME: this is required for dry-v whitelisting
-        stringify!(super)
-      end
-
       def _params
         @result.output.merge(_router_params)
-      end
-
-      def stringify!(result)
-        result.keys.each do |key|
-          value = result.delete(key)
-          result[key.to_s] = case value
-                             when ::Hash
-                               stringify!(value)
-                             when ::Array
-                               value.map(&:to_s)
-                             else
-                               value.to_s
-                             end
-        end
-
-        result
       end
     end
   end