+++ lib/hanami/action/rack.rb
@@ -42,6 +42,9 @@ module Hanami
       # @api private
       HEAD = 'HEAD'.freeze
 
+      # The key that returns router parsed body from the Rack env
+      ROUTER_PARSED_BODY = 'router.parsed_body'.freeze
+
       # Override Ruby's hook for modules.
       # It includes basic Hanami::Action modules to the given class.
       #
@@ -113,16 +116,16 @@ module Hanami
         #   Hanami::Action::Params
         #
         # @api private
-        # @since x.x.x
+        # @since 0.7.0
         def params_class
           @params_class ||= BaseParams
         end
       end
 
-      # @since x.x.x
+      # @since 0.7.0
       # @api private
       module InstanceMethods
-        # @since x.x.x
+        # @since 0.7.0
         # @api private
         def initialize(*)
           super
@@ -200,6 +203,10 @@ module Hanami
       #   end
       def request
         @request ||= ::Hanami::Action::Request.new(@_env)
+      end
+
+      def parsed_request_body
+        @_env.fetch(ROUTER_PARSED_BODY, nil)
       end
 
       private