+++ test/action/format_test.rb
@@ -24,6 +24,16 @@ describe Hanami::Action do
         self.format = params[:format]
       end
     end
+    
+    class Configuration
+      include Hanami::Action
+
+      configuration.default_request_format :jpeg
+
+      def call(params)
+        self.body = format
+      end
+    end
   end
 
   describe '#format' do
@@ -62,6 +72,17 @@ describe Hanami::Action do
 
       @action.format.must_equal    :html
       headers['Content-Type'].must_equal 'text/html; charset=utf-8'
+      status.must_equal                   200
+    end
+
+    # Bug
+    # See https://github.com/hanami/controller/issues/167
+    it "accepts '*/*' and returns configured default format" do
+      action = FormatController::Configuration.new
+      status, headers, _ = action.call({ 'HTTP_ACCEPT' => '*/*' })
+
+      action.format.must_equal    :jpeg
+      headers['Content-Type'].must_equal 'image/jpeg; charset=utf-8'
       status.must_equal                   200
     end